governance/engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dark-factory-policy-engine"
version = "0.1.0"
description = "Deterministic AI governance policy engine for merge decisions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ado = [
    "requests>=2.31.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=5.0",
    "pytest-html>=4.0",
    "hypothesis>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--html=tests/naming-report.html --self-contained-html"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*"]

[tool.setuptools]
py-modules = ["policy_engine", "naming", "naming_data"]

[tool.setuptools.package-data]
"*" = ["schemas/*.json", "policies/*.yaml"]
